scroll to top functin created in geneic function service
