+++ src/app/app.component.html
@@ -1,18 +1,13 @@
-<app-top-bar></app-top-bar>
-<input #input value="Message">
-<button (click)="popup.showAsComponent(input.value)">Show as component</button>
-<div [ngClass]="['container-fluid']" (scroll)="scrollHandler($event)">
+<app-top-bar ></app-top-bar>
+<app-button-top *ngIf="enableScrollToTop" (click)='onActivate($event)'></app-button-top>
+<div [ngClass]="['container-fluid']"  (scroll)="scrollHandler($event)">
   <div [ngClass]="['row']">
     <div [ngClass]="['col-sm-2 col-md-2 col-lg-2']">
-
     </div>
     <div [ngClass]="['col-sm-8 col-md-8 col-lg-8']">
-      <router-outlet></router-outlet>
-      <!-- <button *ngIf='enableScrollToTop' (click)='onActivate($event)'>top</button> -->
-
+      <router-outlet (activate)='onActivate($event)'></router-outlet>
     </div>
     <div [ngClass]="['col-sm-2 col-md-2 col-lg-2']">
-
     </div>
   </div>
