+++ src/app/app.module.ts
@@ -7,8 +7,7 @@ import { AppComponent } from './app.component';
 import { TopBarModule } from './core_modules/top-bar/top-bar.module';
 import { AppRoutingModule } from './app-routing.module';
 import { ButtonTopModule } from './core_modules/button-top/button-top.module';
-import { ButtonTopComponent } from './core_modules/button-top/button-top/button-top.component';
-import { ButtonTopService } from './core_modules/button-top/service-button-top/button-top.service';
+import { GenericFunctionsService } from './shared_modules/generic-functions-service';
 
 
 @NgModule({
@@ -23,9 +22,8 @@ import { ButtonTopService } from './core_modules/button-top/service-button-top/b
     TopBarModule,
     ButtonTopModule
   ],
-  entryComponents: [ButtonTopComponent],
   providers: [
-    ButtonTopService
+    GenericFunctionsService
   ],
   bootstrap: [AppComponent]
 })
