+++ src/app/core_modules/button-top/button-top/button-top.component.html
@@ -1,1 +1,1 @@
-<button type="button" (click)="handleClick()" class="btn btn-default btn-lg">{{label}}</button>
+<a id="return-to-top"><i class="icon-chevron-up"></i></a>
