+++ src/app/core_modules/button-top/button-top/button-top.component.ts
@@ -1,35 +1,10 @@
-import { Component, Input, Output, EventEmitter} from '@angular/core';
-import { animate, state, style, transition, trigger } from '@angular/animations';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-button-top',
   templateUrl: './button-top.component.html',
-  styleUrls: ['./button-top.component.scss'],
-  host: { '[@state]': 'state' },
-  animations: [trigger('state', [
-    state('opened', style({ transform: 'translateY(0%)' })),
-    state('void, closed', style({ transform: 'translateY(100%)', opacity: 0 })),
-    transition('* => *', animate('100ms ease-in')),
-  ])
-  ]
+  styleUrls: ['./button-top.component.scss']
 })
 export class ButtonTopComponent {
-
-  private state: 'opened' | 'closed' = 'closed';
-
   constructor() { }
-
-
-
-  @Input()
-  set message(message: string) {
-    this._message = message;
-    this.state = 'opened';
-  }
-  get message(): string { return this._message; }
-  _message: string;
-
-  @Output()
-  closed = new EventEmitter();
-
 }
