--- src/app/core_modules/button-top/service-button-top/button-top.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ButtonTopService } from './button-top.service';
-
-describe('ButtonTopService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: ButtonTopService = TestBed.get(ButtonTopService);
-    expect(service).toBeTruthy();
-  });
-});
