+++ src/app/shared_modules/rating/rating/rating.component.scss
@@ -1,4 +1,4 @@
-@import url(//netdna.bootstrapcdn.com/font-awesome/3.2.1/css/font-awesome.css);
+
 
 /* Styling h1 and links
 ––––––––––––––––––––––––––––––––– */
