+++ src/styles.css
@@ -1,3 +1,5 @@
+@import url(//netdna.bootstrapcdn.com/font-awesome/3.2.1/css/font-awesome.css);
+
 /* You can add global styles to this file, and also import other style files */
 body {
     padding-top: 4.5rem;
